Fix code review #JC19-352
