+++ apps/vanilla/src/api/index.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { API_HEADER } from '../script/constants';
+export const API_HEADER = 'Api-Key';
 
 export const http = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
