--- libs/core/models/option.ts
@@ -1,19 +1,0 @@
-/** Define the data in one sort option. */
-export interface SortOption {
-
-  /** Title or name of the option. */
-  readonly title: string;
-
-  /** Value of the option. */
-  readonly value: string;
-}
-
-/** Define the data in one order option. */
-export interface OrderOption {
-
-  /** Title or name of the option. */
-  readonly title: string;
-
-  /** Value of the option. */
-  readonly value: string;
-}
