+++ libs/core/models/selectOption.ts
@@ -1,0 +1,9 @@
+/** Define the data in one select option. */
+export interface SelectOption {
+
+  /** Title or name of the option. */
+  readonly title: string;
+
+  /** Value of the option. */
+  readonly value: string;
+}
